clang-tidy: Error on readability-redundant-casting

+++ src/jngl/Rgb.cpp
@@ -37,12 +37,8 @@ void Rgb::setBlue(const float blue) {
 }
 
 Rgb interpolate(Rgb a, Rgb b, float t) {
-	return { static_cast<float>(static_cast<float>(a.getRed()) * (1.f - t) +
-		                        static_cast<float>(b.getRed()) * t),
-		     static_cast<float>(static_cast<float>(a.getGreen()) * (1.f - t) +
-		                        static_cast<float>(b.getGreen()) * t),
-		     static_cast<float>(static_cast<float>(a.getBlue()) * (1.f - t) +
-		                        static_cast<float>(b.getBlue()) * t) };
+	return { a.getRed() * (1.f - t) + b.getRed() * t, a.getGreen() * (1.f - t) + b.getGreen() * t,
+		     a.getBlue() * (1.f - t) + b.getBlue() * t };
 }
 
 } // namespace jngl
